requirements.txt
feedparser>=6.0
httpx>=0.27
python-dotenv>=1.0
google-genai>=1.0
openai>=1.0
